decided that signed and unsigned need to be supported. I wouldn't...
consider satoi as overhead because it processes things before being sent
over to uatoi. The signedness switch is the overhead. It is better to
use unsigned: len * performance(& 255 something) is removed. However, I
think the exchange is good to support the limited amount of signed
numbers it is practical to use.